bank-statement-analyzer/requirements.txt
openai>=1.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
python-dateutil>=2.8.0
rich>=13.0.0
